# applications/oxybank/web/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/deposit": {
      "post": {
        "tags": [
          "Annotation Data Deposit"
        ],
        "summary": "Deposit single QA pair",
        "description": "Store QA pair to annotation platform with type inference",
        "operationId": "deposit_data_api_v1_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DepositResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deposit/batch": {
      "post": {
        "tags": [
          "Annotation Data Deposit"
        ],
        "summary": "Batch deposit",
        "description": "Deposit multiple QA pairs (1-1000 items)",
        "operationId": "deposit_batch_api_v1_deposit_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_DepositBatchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/weather/deposit": {
      "post": {
        "tags": [
          "KB:weather",
          "KB:weather"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base weather. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_weather_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_weather",
                "description": "Data to write (single item). Fields: city_key, date, max_temperature, min_temperature, weather, wind_force_direction, sunshine_duration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/weather-temp/deposit": {
      "post": {
        "tags": [
          "KB:weather-temp",
          "KB:weather-temp"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base weather-temp. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_weather_temp_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_weather-temp",
                "description": "Data to write (single item). Fields: city_key, date, max_temperature, min_temperature, weather, wind_force_direction, sunshine_duration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/unstructure_doc_test/deposit": {
      "post": {
        "tags": [
          "KB:unstructure_doc_test",
          "KB:unstructure_doc_test"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base unstructure_doc_test. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_unstructure_doc_test_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_unstructure_doc_test",
                "description": "Data to write (single item). Fields: chunk_to_return, chunk_to_emb"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/qa_bank_2/deposit": {
      "post": {
        "tags": [
          "KB:qa_bank_2",
          "KB:qa_bank_2"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base qa_bank_2. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_qa_bank_2_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_qa_bank_2",
                "description": "Data to write (single item). Fields: question, answer, score, caller, callee, remark, source_trace_id, source_request_id, data_type, priority, category"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/anno_bank/deposit": {
      "post": {
        "tags": [
          "KB:anno_bank",
          "KB:anno_bank"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base anno_bank. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_anno_bank_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_anno_bank",
                "description": "Data to write (single item). Fields: task_id, type, priority, status, create_time, update_time, query, answer, is_satisfied, reason"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/testK/deposit": {
      "post": {
        "tags": [
          "KB:testK",
          "KB:testK"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base testK. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_testK_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_testK",
                "description": "Data to write (single item). Fields: chunk_to_return, chunk_to_emb"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/hyy_new_structure_1/deposit": {
      "post": {
        "tags": [
          "KB:hyy_new_structure_1",
          "KB:hyy_new_structure_1"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base hyy_new_structure_1. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_hyy_new_structure_1_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_hyy_new_structure_1",
                "description": "Data to write (single item). Fields: name, age, score"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/kb_un/deposit": {
      "post": {
        "tags": [
          "KB:kb_un",
          "KB:kb_un"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base kb_un. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_kb_un_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_kb_un",
                "description": "Data to write (single item). Fields: chunk_to_return, chunk_to_emb"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/test222/deposit": {
      "post": {
        "tags": [
          "KB:test222",
          "KB:test222"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base test222. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_test222_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_test222",
                "description": "Data to write (single item). Fields: name, age, score"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/hyy_demo_1/deposit": {
      "post": {
        "tags": [
          "KB:hyy_demo_1",
          "KB:hyy_demo_1"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base hyy_demo_1. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_hyy_demo_1_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_hyy_demo_1",
                "description": "Data to write (single item). Fields: name, age, score"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/hyy_test_111/deposit": {
      "post": {
        "tags": [
          "KB:hyy_test_111",
          "KB:hyy_test_111"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base hyy_test_111. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_hyy_test_111_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_hyy_test_111",
                "description": "Data to write (single item). Fields: name, age, score, tag"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/testK2/deposit": {
      "post": {
        "tags": [
          "KB:testK2",
          "KB:testK2"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base testK2. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_testK2_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_testK2",
                "description": "Data to write (single item). Fields: chunk_to_return, chunk_to_emb"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/test3/deposit": {
      "post": {
        "tags": [
          "KB:test3",
          "KB:test3"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base test3. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_test3_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_test3",
                "description": "Data to write (single item). Fields: name, age, score"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/hyyaaaaaaaaaaa/deposit": {
      "post": {
        "tags": [
          "KB:hyyaaaaaaaaaaa",
          "KB:hyyaaaaaaaaaaa"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base hyyaaaaaaaaaaa. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_hyyaaaaaaaaaaa_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_hyyaaaaaaaaaaa",
                "description": "Data to write (single item). Fields: name, age, score"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/kb/daefwasfa1111/deposit": {
      "post": {
        "tags": [
          "KB:daefwasfa1111",
          "KB:daefwasfa1111"
        ],
        "summary": "Data write interface",
        "description": "Write data to knowledge base daefwasfa1111. Accepts a single data item matching the knowledge base schema.",
        "operationId": "deposit_data_kb_daefwasfa1111_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDataItem_daefwasfa1111",
                "description": "Data to write (single item). Fields: chunk_to_return, chunk_to_emb"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_DepositBatchResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DepositBatchResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[DepositBatchResponse]"
      },
      "APIResponse_DepositResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DepositResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[DepositResponse]"
      },
...
      "DepositBatchRequest": {
        "properties": {
          "data_list": {
            "items": {
              "$ref": "#/components/schemas/DepositRequest"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Data List",
            "description": "Data list"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id",
            "description": "Batch ID (auto-generated if not provided)"
          }
        },
        "type": "object",
        "required": [
          "data_list"
        ],
        "title": "DepositBatchRequest",
        "description": "Batch data deposit request"
      },
...
      "DepositDataItem_testK": {
        "properties": {
          "chunk_to_return": {
            "type": "string",
            "title": "Chunk To Return",
            "description": "chunk_to_return"
          },
          "chunk_to_emb": {
            "type": "string",
            "title": "Chunk To Emb",
            "description": "chunk_to_emb"
          }
        },
        "type": "object",
        "required": [
          "chunk_to_return",
          "chunk_to_emb"
        ],
        "title": "DepositDataItem_testK"
      },
...
      "DepositDataItem_unstructure_doc_test": {
        "properties": {
          "chunk_to_return": {
            "type": "string",
            "title": "Chunk To Return",
            "description": "chunk_to_return"
          },
          "chunk_to_emb": {
            "type": "string",
            "title": "Chunk To Emb",
            "description": "chunk_to_emb"
          }
        },
        "type": "object",
        "required": [
          "chunk_to_return",
          "chunk_to_emb"
        ],
        "title": "DepositDataItem_unstructure_doc_test"
      },
      "DepositDataItem_weather": {
        "properties": {
          "city_key": {
            "type": "string",
            "title": "City Key",
            "description": "city_key"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "description": "date"
          },
          "max_temperature": {
            "type": "number",
            "title": "Max Temperature",
            "description": "max_temperature"
          },
          "min_temperature": {
            "type": "number",
            "title": "Min Temperature",
            "description": "min_temperature"
          },
          "weather": {
            "type": "string",
            "title": "Weather",
            "description": "weather"
          },
          "wind_force_direction": {
            "type": "string",
            "title": "Wind Force Direction",
            "description": "wind_force_direction"
          },
          "sunshine_duration": {
            "type": "number",
            "title": "Sunshine Duration",
            "description": "sunshine_duration"
          }
        },
        "type": "object",
        "required": [
          "city_key",
          "date",
          "max_temperature",
          "min_temperature",
          "weather",
          "wind_force_direction",
          "sunshine_duration"
        ],
        "title": "DepositDataItem_weather"
      },
      "DepositDataItem_weather-temp": {
        "properties": {
          "city_key": {
            "type": "string",
            "title": "City Key",
            "description": "city_key"
          },
          "date": {
            "type": "string",
            "title": "Date",
            "description": "date"
          },
          "max_temperature": {
            "type": "number",
            "title": "Max Temperature",
            "description": "max_temperature"
          },
          "min_temperature": {
            "type": "number",
            "title": "Min Temperature",
            "description": "min_temperature"
          },
          "weather": {
            "type": "string",
            "title": "Weather",
            "description": "weather"
          },
          "wind_force_direction": {
            "type": "string",
            "title": "Wind Force Direction",
            "description": "wind_force_direction"
          },
          "sunshine_duration": {
            "type": "number",
            "title": "Sunshine Duration",
            "description": "sunshine_duration"
          }
        },
        "type": "object",
        "required": [
          "city_key",
          "date",
          "max_temperature",
          "min_temperature",
          "weather",
          "wind_force_direction",
          "sunshine_duration"
        ],
        "title": "DepositDataItem_weather-temp"
      },
      "DepositRequest": {
        "properties": {
          "question": {
            "type": "string",
            "minLength": 2,
            "title": "Question",
            "description": "Question content"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Answer content"
          },
          "source_trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Trace Id",
            "description": "Original trace_id"
          },
          "source_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Request Id",
            "description": "Original request_id"
          },
          "source_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Group Id",
            "description": "Session group ID"
          },
          "caller": {
            "type": "string",
            "title": "Caller",
            "description": "Caller",
            "default": ""
          },
          "callee": {
            "type": "string",
            "title": "Callee",
            "description": "Callee",
            "default": ""
          },
          "caller_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller Type",
            "description": "Caller type"
          },
          "callee_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callee Type",
            "description": "Callee type"
          },
          "data_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Type",
            "description": "Data type (auto-inferred if not provided)"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4,
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "description": "Priority (auto-set if not provided)"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "Data category"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Data tags"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra",
            "description": "Extra data"
          }
        },
        "type": "object",
        "required": [
          "question",
          "answer"
        ],
        "title": "DepositRequest",
        "description": "Single data deposit request"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}